// apps/web/src/api/generated/organizer-tournaments/organizerRegistrationListItemDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CLUTCHA API
 * API documentation for the CLUTCHA online and on-site esports tournament platform.
 * OpenAPI spec version: 1.0.0
 */
import type { OrganizerRegistrationEligibilityDto } from './organizerRegistrationEligibilityDto';
import type { OrganizerRegistrationListItemDtoApprovalStatus } from './organizerRegistrationListItemDtoApprovalStatus';
import type { OrganizerRegistrationListItemDtoPaymentStatus } from './organizerRegistrationListItemDtoPaymentStatus';
import type { OrganizerRegistrationListItemDtoStatus } from './organizerRegistrationListItemDtoStatus';
import type { OrganizerRegistrationTeamDto } from './organizerRegistrationTeamDto';

export interface OrganizerRegistrationListItemDto {
  registrationId: string;
  team: OrganizerRegistrationTeamDto;
  status: OrganizerRegistrationListItemDtoStatus;
  paymentStatus: OrganizerRegistrationListItemDtoPaymentStatus;
  approvalStatus: OrganizerRegistrationListItemDtoApprovalStatus;
  rulesVersion: string;
  submittedAt: string;
  /** @nullable */
  rejectionReason?: string | null;
  eligibility: OrganizerRegistrationEligibilityDto;
}